+++ inertia/ssr.ts
@@ -2,6 +2,7 @@ import { client } from '~/client'
 import Layout from '~/layouts/default.vue'
 import { createInertiaApp } from '@inertiajs/vue3'
 import { TuyauProvider } from '@adonisjs/inertia/vue'
+import { MotionPlugin } from '@vueuse/motion'
 import { renderToString } from '@vue/server-renderer'
 import { createSSRApp, h, type DefineComponent } from 'vue'
 import { resolvePageComponent } from '@adonisjs/inertia/helpers'
@@ -20,7 +21,9 @@ export default function render(page: any) {
     setup: ({ App, props, plugin }) => {
       return createSSRApp({
         render: () => h(TuyauProvider, { client }, { default: () => h(App, props) }),
-      }).use(plugin)
+      })
+        .use(plugin)
+        .use(MotionPlugin)
     },
   })
 }